Add compression to the format
We make the frame structure more flexible. The codec can store either compressed or raw frames. A compressed frame works by defining a set of colors that is used in the frame and then referencing them with indices. This reduces the size of each pixel from 3 bytes (rgb) to 1 byte (index pointing to 1 of max 255 colors).

Encoding the example file (hello world) using just raw frames results in ~105kB of data, whereas enabling the compression mode results in ~55kB of data.