Run-length encoding and bitshift for quantization
Replace the weird home-made quantization math with more standard bitshift. 1 bit is reserved as a flag to control run-length-encoding. If the exact same (quantized) pixel is repeated over and over, we encode that with a single byte that tells the decoder the number of consecutive identical pixels (up to 127 as that's what we fit in 7 bits).